Better timing of removing old connections.
Don't remove it right away on login1. Do it after login2 is successful.
Login2 is successful when the person answer's the challenge correctly.
That means you're 100% sure the old connections aren't needed because
the replacements are right there.

Removing old connections on login1 allows you to arbitrarily kick out
anybody you feel like by requesting a challenge and ignoring it.